FooterBottom is refactor
